+++ R/oncoplot.R
@@ -25,6 +25,7 @@
 #' @param annotationDat If MAF file was read without clinical data, provide a custom \code{data.frame} with a column \code{Tumor_Sample_Barcode} containing sample names along with rest of columns with annotations.
 #' You can specify which columns to be drawn using `clinicalFeatures` argument.
 #' @param pathways Default `NULL`. Can be `auto`, or a two column data.frame/tsv-file with genes and correspoding pathway mappings.`
+#' @param path_order Default `NULL` Manually specify the order of pathways
 #' @param selectedPathways Manually provide the subset of pathway names to be seletced from `pathways`. Default NULL. In case `pathways` is `auto` draws top 3 altered pathways.
 #' @param draw_titv logical Includes TiTv plot. \code{FALSE}
 #' @param showTumorSampleBarcodes logical to include sample names.
@@ -89,7 +90,7 @@ oncoplot = oncoplot = function(maf, top = 20, minMut = NULL, genes = NULL, alter
                                rightBarData = NULL, rightBarLims = NULL,
                                topBarData = NULL, logColBar = FALSE, includeColBarCN = TRUE,
                                clinicalFeatures = NULL, annotationColor = NULL, annotationDat = NULL,
-                               pathways = NULL, selectedPathways = NULL, draw_titv = FALSE,
+                               pathways = NULL, path_order = NULL, selectedPathways = NULL, draw_titv = FALSE,
                                showTumorSampleBarcodes = FALSE, barcode_mar = 4, barcodeSrt = 90, gene_mar = 5,
                                anno_height = 1, legend_height = 4,
                                sortByAnnotation = FALSE, groupAnnotationBySize = TRUE, annotationOrder = NULL,
@@ -120,9 +121,13 @@ oncoplot = oncoplot = function(maf, top = 20, minMut = NULL, genes = NULL, alter
   } else if(!is.null(pathways)){ #If user provides pathway list
 
     if(is(pathways, 'data.frame')){
+      pathways = data.table::copy(pathways)
       colnames(pathways)[1:2] = c('Gene', 'Pathway')
       data.table::setDT(x = pathways)
       pathways = pathways[!duplicated(Gene)][,.(Gene, Pathway)]
+      if(!is.null(selectedPathways)){
+        pathways = pathways[Pathway %in% selectedPathways]
+      }
     }else if(file.exists(pathways)){
       pathways = data.table::fread(file = pathways)
       colnames(pathways)[1:2] = c('Gene', 'Pathway')
@@ -330,6 +335,14 @@ oncoplot = oncoplot = function(maf, top = 20, minMut = NULL, genes = NULL, alter
       temp_dat = temp_dat[temp_dat_ord]
     }
 
+    if(!is.null(path_order)){
+      path_order = intersect(x = path_order, names(temp_dat) )
+      temp_dat = temp_dat[path_order]
+      if(length(temp_dat) == 0){
+        stop("None of the pathways match the provided order!")
+      }
+    }
+
     nm = lapply(seq_along(temp_dat), function(i){
       x = numMat[temp_dat[[i]]$Gene,, drop = FALSE]
       x = rbind(x, apply(x, 2, function(y) ifelse(test = sum(y) == 0, yes = 0, no = 99)))
@@ -417,6 +430,8 @@ oncoplot = oncoplot = function(maf, top = 20, minMut = NULL, genes = NULL, alter
     }
     if(logColBar){
       mtext(text = "(log10)", side = 2, line = 2, cex = 0.6)
+    }else{
+      mtext(text = "TMB", side = 2, line = 2, cex = 0.6)
     }
   }else if(!is.null(topBarData) & drawColBar){
     # Draw extra clinical data in top
@@ -492,8 +507,10 @@ oncoplot = oncoplot = function(maf, top = 20, minMut = NULL, genes = NULL, alter
 
     if(drawColBar){
       par(mar = c(0.25 , 0, 1, 2), xpd = TRUE)
+
       plot(x = NA, y = NA, type = "n", axes = FALSE,
            xlim = side_bar_lims, ylim = c(0, 1), xaxs = "i")
+      rightBarTitle = "No. of samples"
     }
   }
 
@@ -793,6 +810,7 @@ oncoplot = oncoplot = function(maf, top = 20, minMut = NULL, genes = NULL, alter
         }
       }
       axis(side = 3, at = side_bar_lims, outer = FALSE, line = 0.25)
+      mtext(text = rightBarTitle, side = 3, line = 0.5, cex = 0.6)
     }else{
 
       plot(x = NA, y = NA, type = "n", xlim = side_bar_lims, ylim = c(0, nrow(rightBarData)),
@@ -972,9 +990,29 @@ oncoplot = oncoplot = function(maf, top = 20, minMut = NULL, genes = NULL, alter
   leg_classes = vc_col[vc_codes[2:length(vc_codes)]]
   leg_classes_pch = rep(15, length(leg_classes))
   if(additionalFeature_legend){
-    leg_classes = c(leg_classes,"gray70")
-    names(leg_classes)[length(leg_classes)] = paste(additionalFeature, collapse = ":")
-    leg_classes_pch = c(leg_classes_pch, additionalFeaturePch)
+
+    if(!is(object = additionalFeature, class2 = "list")){
+      if(length(additionalFeature) < 2){
+        stop("additionalFeature must be of length two. See ?oncoplot for details.")
+      }else{
+        additionalFeature = list(additionalFeature)
+      }
+    }
+
+    if(length(additionalFeaturePch) != length(additionalFeature)){
+      additionalFeaturePch = rep(additionalFeaturePch, length(additionalFeature))
+    }
+
+    if(length(additionalFeatureCol) != length(additionalFeature)){
+      additionalFeatureCol = rep(additionalFeatureCol, length(additionalFeature))
+    }
+
+    for(af_idx in 1:length(additionalFeature)){
+      af = additionalFeature[[af_idx]]
+      leg_classes = c(leg_classes,additionalFeatureCol[af_idx])
+      names(leg_classes)[length(leg_classes)] = paste(af, collapse = ":")
+      leg_classes_pch = c(leg_classes_pch, additionalFeaturePch[af_idx])
+    }
   }
 
   lep = legend("topleft", legend = names(leg_classes),