+++ R/subsetMaf.R
@@ -168,11 +168,13 @@ subsetMaf = function(maf, tsb = NULL, genes = NULL, query = NULL, clinQuery = NU
       stop("Subsetting has resulted in zero non-synonymous variants!")
     }
 
-    mafSummary = summarizeMaf(maf.dat, chatty = FALSE, anno = maf.anno)
+    m = MAF(nonSyn = maf.dat, syn = maf.silent, clinicalData = maf.anno)
 
-    m = MAF(data = maf.dat, variants.per.sample = mafSummary$variants.per.sample, variant.type.summary = mafSummary$variant.type.summary,
-              variant.classification.summary = mafSummary$variant.classification.summary, gene.summary = mafSummary$gene.summary,
-              summary = mafSummary$summary, maf.silent = maf.silent, clinical.data = droplevels(mafSummary$sample.anno))
+    # mafSummary = summarizeMaf(maf.dat, chatty = FALSE, anno = maf.anno)
+    #
+    # m = MAF(data = maf.dat, variants.per.sample = mafSummary$variants.per.sample, variant.type.summary = mafSummary$variant.type.summary,
+    #           variant.classification.summary = mafSummary$variant.classification.summary, gene.summary = mafSummary$gene.summary,
+    #           summary = mafSummary$summary, maf.silent = maf.silent, clinical.data = droplevels(mafSummary$sample.anno))
 
     return(m)
   }else{